back/pyproject.toml
[project]
name = "back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "channels>=4.3.2",
    "daphne>=4.2.1",
    "django>=6.0.3",
    "django-extensions>=4.1",
    "django-ninja>=1.6.2",
]
